Minor restructuring, prepare to add rendering
