+++ ClientApp/src/app/components/auth/register/register.component.ts
@@ -6,10 +6,10 @@ import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatIconModule } from "@angular/material/icon";
 import { MatInputModule } from "@angular/material/input";
 import { RouterModule } from "@angular/router";
-import { ControlsOf } from "../../../helpers";
-import { RegisterRequest } from "../../../models";
-import { IdentityService } from "../../../services";
 import { AuthCardComponent } from "../auth-card/auth-card.component";
+import { ControlsOf } from "../../../helpers/controls-of.type";
+import { RegisterRequest } from "../../../models/register-request.model";
+import { IdentityService } from "../../../services/identity.service";
 
 
 @Component({
