+++ ClientApp/src/app/components/game/game-screen/game-screen.component.html
@@ -1,1 +1,1 @@
-<p>game-screen works!</p>
+<canvas #gameCanvas class="game-canvas" (click)="onCanvasClick($event)">
