+++ ClientApp/src/app/components/game/game-screen/game-screen.component.scss
@@ -3,3 +3,8 @@
   flex-direction: column;
   flex: 1;
 }
+
+.game-canvas {
+  width: 100%;
+  height: 100%;
+}
