+++ ClientApp/src/app/components/game/game/game.component.scss
@@ -46,5 +46,4 @@ app-game-bar-quick-access {
 
 app-game-screen {
   top: calc(var(--map-select-height)*-1);
-  background-color: black;
 }
