+++ ClientApp/src/app/components/game/world-object-list/world-object-list.component.ts
@@ -4,11 +4,11 @@ import { Component, ChangeDetectionStrategy, inject } from "@angular/core";
 import { MatIconModule } from "@angular/material/icon";
 import { MatTreeModule } from "@angular/material/tree";
 import { map } from "rxjs";
-import { WorldObjectResponse } from "../../../models";
 import { TreeService, NamedTreeNode } from "../../../services/tree-service";
 import { WorldObjectService } from "../../../services/world-object.service";
 import { MatButtonModule } from "@angular/material/button";
 import { WorldObjectComponent } from "../world-object/world-object.component";
+import { WorldObjectResponse } from "../../../models/world-object-response";
 
 @Component({
   selector: 'app-world-object-list',
