--- ClientApp/src/app/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from './controls-of.type';
-export * from './path-to-name.pipe'
