--- ClientApp/src/app/models/index.ts
@@ -1,7 +1,0 @@
-export * from './login-request.model';
-export * from './register-request.model';
-export * from './token-response.model';
-export * from './create-world-request';
-export * from './edit-world-request';
-export * from './world-response';
-export * from './world-object-response';
