--- ClientApp/src/app/services/index.ts
@@ -1,3 +1,0 @@
-export * from './api.service';
-export * from './identity.service';
-export * from './world.service';
